+++ frontend/e2e/text-annotator.test.ts
@@ -1,7 +1,7 @@
 import { expect, test } from "@playwright/test";
 import { log } from "../playwright.config";
 /** https://github.com/microsoft/playwright/issues/23662 */
-import example from "../src/pages/explore/text-annotator.json";
+import example from "../src/data/text-annotator.json";
 
 log();
