+++ frontend/src/pages/about/PageOurStory.vue
