--- frontend/src/pages/about/PageOutreach.vue
@@ -1,85 +1,0 @@
-<!-- 
-    This is a page for information about outreach and community engagement.
- -->
-
-<template>
-  <AppSection>
-    <AppHeading>Outreach</AppHeading>
-    <p>
-      The Monarch Initiative is a community effort, and we welcome your
-      participation!
-    </p>
-
-    <AppFlex>
-      <AppButton to="/help" icon="handshake-angle" text="Get Help" />
-      <AppButton to="/#follow" icon="share-nodes" text="Follow" />
-    </AppFlex>
-  </AppSection>
-
-  <AppSection>
-    <AppHeading>Ways to Get Involved</AppHeading>
-
-    <AppLink to="https://mondo.monarchinitiative.org"
-      >The Mondo Disease Ontology (Mondo)</AppLink
-    >
-
-    <p>
-      Mondo aims to harmonize disease definitions across the world. The name
-      Mondo comes from the latin word ‘mundus’ and means ‘for the world.’
-    </p>
-
-    <ul>
-      <li>
-        The Mondo Outreach calls aim to bring together stakeholders and clinical
-        terminology experts to discuss specific use cases and requirements. Our
-        goal is to improve Mondo with a focus on our user needs.
-      </li>
-      <ul>
-        <li>
-          These calls are scheduled monthly. You can find more information about
-          the calls
-          <AppLink
-            to="https://mondo.monarchinitiative.org/pages/workshop/#outreach"
-            >here</AppLink
-          >.
-        </li>
-      </ul>
-      <li>
-        Join the Mondo users mailing list
-        <AppLink to="https://groups.google.com/g/mondo-users"> here </AppLink>
-      </li>
-    </ul>
-
-    <hr />
-
-    <AppLink
-      to="https://oboacademy.github.io/obook/courses/monarch-obo-training/"
-      >OBO Academy: Monarch Training Series</AppLink
-    >
-    <p>
-      Training for Semantic Engineers, Ontology Curators, and Knowledge
-      Scientists on basic technologies in and around the Monarch Initiative and
-      Open Biological and Biomedical Ontologies (OBO).
-    </p>
-
-    <ul>
-      <li>
-        The OBO Academy develops self-paced tutorials in the OBOOK, hosts
-        fortnightly live training sessions on Zoom, and maintains a YouTube
-        channel with lesson recordings.
-      </li>
-      <li>
-        YouTube channel:
-        <AppLink to="https://www.youtube.com/@obo-academy"
-          >https://www.youtube.com/@obo-academy</AppLink
-        >
-      </li>
-      <li>
-        Join the OBO Academy mailing list
-        <AppLink to="https://groups.google.com/g/monarch-tutorials">
-          here
-        </AppLink>
-      </li>
-    </ul>
-  </AppSection>
-</template>
