+++ frontend/src/pages/about/PageTeam.vue
@@ -1,104 +1,98 @@
-<!--
-  about team page
-
-  all monarch-related team members and info about them
--->
-
 <template>
-  <!-- brief intro -->
-  <AppSection>
-    <AppHeading>Team</AppHeading>
-    <p>
-      Monarch is made possible thanks to a diverse and dedicated team of
-      biologists, scientists, and programmers from various schools and
-      institutes:
-    </p>
-    <div class="groups">
-      <AppLink
-        v-for="(group, index) in team"
-        :key="index"
-        :to="'#' + kebabCase(group.name)"
-        :replace="true"
-        >{{ group.name }}</AppLink
-      >
-    </div>
-  </AppSection>
+  <AppBreadcrumb />
+  <PageTitle id="about-team" title="Meet the team" />
+  <AppSection width="big">
+    <!-- Page Title -->
+    <section class="university-group">
+      <p>
+        Monarch is made possible thanks to a diverse and dedicated team of
+        biologists, scientists, and programmers from various schools and
+        institutes:
+      </p>
 
-  <!-- list each group -->
-  <AppSection v-for="(group, groupIndex) in team" :key="groupIndex" width="big">
-    <AppHeading>
-      {{ group.name }}
-    </AppHeading>
-    <AppGroup :name="group.name" :link="group.link" />
-    <AppGallery :cols="group.name.includes('Alumni') ? 6 : 4">
-      <template v-if="!group.type">
-        <AppMember
-          v-for="(member, memberIndex) in group.members"
-          :key="memberIndex"
-          :name="member.name"
-          :role="'role' in member ? member.role : ''"
-          :link="'link' in member ? member.link : ''"
-          :no-image="group.name.includes('Alumni') ? true : undefined"
-        />
-      </template>
-      <template v-else>
+      <div class="groups-container">
+        <div v-for="i in 2" :key="i" class="group-box">
+          <ul>
+            <li
+              v-for="(group, index) in team.slice(
+                (i - 1) * (team.length / 2),
+                i * (team.length / 2),
+              )"
+              :key="index"
+            >
+              <AppLink :to="'#' + kebabCase(group.name)" :replace="true">
+                {{ group.name }}
+              </AppLink>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </section>
+
+    <section
+      v-for="(group, groupIndex) in team"
+      :id="kebabCase(group.name)"
+      :key="groupIndex"
+      class="university-block"
+    >
+      <div class="university-header">{{ group.name }}</div>
+      <div class="group-logo-wrapper">
         <AppGroup
-          v-for="(member, memberIndex) in group.members"
-          :key="memberIndex"
-          :name="member.name"
-          :link="member.link"
+          v-if="group.link"
+          :name="group.name"
+          :link="group.link"
+          class="university-logo"
         />
-      </template>
-    </AppGallery>
-  </AppSection>
+      </div>
 
-  <!-- funding sources -->
-  <AppSection>
-    <AppHeading>Funding</AppHeading>
-    <ul>
-      <li>
-        OFFICE OF THE DIRECTOR, NATIONAL INSTITUTES OF HEALTH
-        <br />
-        <AppLink to="https://reporter.nih.gov/project-details/10173498">
-          The Monarch Initiative: Linking diseases to model organism
-          resources</AppLink
-        >
-        <br />
-        2R24OD011883-10A1
-      </li>
-      <li>
-        NATIONAL HUMAN GENOME RESEARCH INSTITUTE, Center of Excellence in Genome
-        Sciences
-        <br />
-        <AppLink to="https://reporter.nih.gov/project-details/10448140">
-          A phenomics-first resource for interpretation of variants
-        </AppLink>
-        <br />
-        7RM1HG010860-02
-      </li>
-      <li>
-        NATIONAL HUMAN GENOME RESEARCH INSTITUTE
-        <br />
-        <AppLink to="https://reporter.nih.gov/project-details/10269338">
-          The Human Phenotype Ontology: Accelerating Computational Integration
-          of Clinical Data for Genomics</AppLink
-        >
-        <br />
-        1U24HG011449-01A1
-      </li>
-    </ul>
+      <div class="members">
+        <AppGallery :cols="group.name.includes('Alumni') ? 6 : 4">
+          <template v-if="!group.type">
+            <AppMember
+              v-for="(member, memberIndex) in group.members"
+              :key="memberIndex"
+              :name="member.name"
+              :role="'role' in member ? member.role : ''"
+              :link="'link' in member ? member.link : ''"
+              :no-image="group.name.includes('Alumni') ? true : undefined"
+            />
+          </template>
+          <template v-else>
+            <AppGroup
+              v-for="(member, memberIndex) in group.members"
+              :key="memberIndex"
+              :name="member.name"
+              :link="member.link"
+            />
+          </template>
+        </AppGallery>
+      </div>
+    </section>
   </AppSection>
 </template>
-
 <script setup lang="ts">
 import { kebabCase } from "lodash";
+import AppBreadcrumb from "@/components/AppBreadcrumb.vue";
 import AppGroup from "@/components/AppGroup.vue";
 import AppMember from "@/components/AppMember.vue";
-import team from "./team.json";
+import AppSection from "@/components/AppSection.vue";
+import PageTitle from "@/components/ThePageTitle.vue";
+import team from "@/data/team.json";
 </script>
+<style scoped lang="scss">
+$wrap: 1000px;
+html {
+  scroll-behavior: smooth;
+}
 
-<style lang="scss" scoped>
-$wrap: 800px;
+:root {
+  scroll-padding-top: 90px;
+}
+.university-group {
+  display: flex;
+  flex-direction: column;
+  gap: 2em;
+}
 
 .groups {
   display: grid;
@@ -106,6 +100,7 @@ $wrap: 800px;
   justify-items: flex-start;
   width: 100%;
   gap: 10px;
+
   text-align: left;
 }
 
@@ -113,5 +108,79 @@ $wrap: 800px;
   .groups {
     grid-template-columns: 1fr;
   }
+}
+
+.groups-container {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+
+  gap: 3em;
+  @media (max-width: $wrap) {
+    gap: 1.5em;
+  }
+}
+
+.group-box {
+  width: 100%;
+  max-width: 400px;
+  padding: 1rem 1.5rem;
+  border-radius: 8px;
+  background-color: rgba(166, 236, 242, 0.3411764706);
+  ul {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+    white-space: nowrap;
+
+    li + li {
+      margin-top: 0.5rem;
+    }
+
+    a {
+      color: #111;
+      font-weight: normal;
+      text-decoration: underline;
+    }
+  }
+}
+
+@media (min-width: $wrap) {
+  .group-box {
+    flex: 1;
+  }
+}
+
+.university-block {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  gap: 0.3em;
+}
+
+.university-header {
+  width: 100%;
+  padding: 0.5em;
+  background-color: rgba(166, 236, 242, 0.3411764706);
+  font-weight: 500;
+  font-size: 1.5rem;
+  text-align: center;
+}
+
+.university-logo {
+  display: block;
+  max-height: 60px;
+  margin: 1.5rem auto;
+}
+.group-logo-wrapper {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  min-height: 2em;
+}
+.members {
+  display: flex;
+  justify-content: center;
 }
 </style>
