--- frontend/src/pages/about/PageTerms.vue
@@ -1,138 +1,0 @@
-<!--
-  about terms page
-
-  agreements and legal stuff
--->
-
-<template>
-  <AppSection>
-    <AppHeading>Terms of Use</AppHeading>
-  </AppSection>
-
-  <AppSection>
-    <AppHeading>License</AppHeading>
-    <p>
-      This web application and its source code are licensed under the
-      <AppLink to="https://opensource.org/licenses/BSD-3-Clause"
-        >3-Clause BSD License</AppLink
-      >. Licensing information for the data sources and ontologies that make up
-      the Monarch knowledge graph can be found on the
-      <AppLink to="https://monarch-initiative.github.io/monarch-ingest/Sources"
-        >sources page</AppLink
-      >. Licensing information for tools, APIs, algorithms, and workflows can be
-      found on the <AppLink to="/overview">overview page</AppLink>.
-    </p>
-    <AppHeading>Tracking</AppHeading>
-    <p>
-      This web application tracks user behavior such as navigation and clicks
-      through the use of HotJar and Google Analytics. These background
-      scripts/services take significant steps to remove identifying information
-      from the data they collect to ensure anonymity. The Monarch Initiative
-      pledges to never inspect or disclose any private, identifying information
-      collected, and to only use it for the purposes of improving Monarch. To
-      learn more about how Hotjar maintains privacy and complies with relevant
-      privacy laws, see
-      <AppLink to="https://www.hotjar.com/legal/compliance/gdpr-commitment/"
-        >their explanation here</AppLink
-      >.
-    </p>
-    <AppHeading>Local Data</AppHeading>
-    <p>
-      For the convenience of the user on repeat visits to the site, this web
-      application also saves certain information locally, such as recent
-      searches and form drafts. This data only exists locally on the user's
-      device, and is not sent to the Monarch team or any other party.
-    </p>
-  </AppSection>
-
-  <AppSection>
-    <AppHeading>Definitions</AppHeading>
-    <p>
-      The first-party software and resources used and referenced on this website
-      &mdash; including but not limited to the Monarch Website (this website
-      itself), BioLink, and Exomiser &mdash; are hereby referred to as the
-      "Monarch Tools". The associated
-      <AppLink to="/team">Universities, Institutions, and other groups</AppLink>
-      that develop and maintain the Monarch Tools are hereby collectively
-      refered to as the "Monarch Initiative".
-    </p>
-    <AppHeading>Intended use</AppHeading>
-    <p>
-      Monarch Tools should not be used for direct diagnostic use or medical
-      decision-making. If you have any health questions about the information
-      contained on this website, please contact a medical professional.
-    </p>
-    <AppHeading>Disclaimer</AppHeading>
-    <p>
-      YOU ACKNOWLEDGE THAT THE MONARCH TOOLS, INCLUDING WITHOUT LIMITATION, THE
-      DATABASE OF INFORMATION CONTAINED THEREIN, IS EXPERIMENTAL AND ACADEMIC IN
-      NATURE, AND IS NOT LICENSED BY THE U.S. FOOD AND DRUG ADMINISTRATION OR
-      ANY OTHER REGULATORY BODY. THE MONARCH TOOLS ARE PROVIDED ON AN "AS IS"
-      AND “AS AVAILABLE” BASIS WITH ALL FAULTS, WITHOUT WARRANTY OR
-      REPRESENTATION OF ANY KIND, EXPRESS OR IMPLIED. THE MONARCH INITIATIVE
-      DISCLAIMS ALL WARRANTIES AND REPRESENTATIONS, EXPRESS OR IMPLIED, WITH
-      RESPECT TO THE MONARCH TOOLS, INCLUDING WITHOUT LIMITATION, ANY WARRANTIES
-      OF TITLE, MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
-      NON-INFRINGEMENT OF THIRD PARTY RIGHTS. THIS DISCLAIMER APPLIES TO THE
-      FULLEST EXTENT POSSIBLE. THE MONARCH INITIATIVE DOES NOT WARRANT THAT THE
-      MONARCH TOOLS WILL MEET YOUR REQUIREMENTS OR THAT THE USE OF THE MONARCH
-      TOOLS WILL BE FREE OF INFECTION OR VIRUSES, ERROR-FREE, UNINTERRUPTED,
-      SECURE OR WILL PRODUCE ACCURATE RESULTS. YOU SHALL BEAR TOTAL AND
-      EXCLUSIVE RESPONSIBILITY AND RISK FOR THE USE OF THE MONARCH TOOLS. YOU
-      SHALL BE SOLELY RESPONSIBLE FOR ANY RESULTING DAMAGE TO YOUR COMPUTER
-      SYSTEMS OR LOSS OF DATA. IF YOU ARE DISSATISFIED WITH THE MONARCH TOOLS,
-      OR WITH ANY OF THESE TERMS, YOUR SOLE AND EXCLUSIVE REMEDY IS TO
-      DISCONTINUE USING THE MONARCH TOOLS.
-    </p>
-    <AppHeading>Limitation of liability</AppHeading>
-    <p>
-      IN NO EVENT SHALL THE MONARCH INITIATIVE, OR THEIR RESPECTIVE DIRECTORS,
-      OFFICERS, EMPLOYEES, AGENTS, LICENSEES, LICENSORS, CONTRIBUTORS, FOUNDERS
-      OR ADVISORS, OR ANY OTHER PERSON WHO HAS BEEN INVOLVED IN THE CREATION OR
-      PRODUCTION OF THE MONARCH TOOLS (EACH A “RELEASED PARTY” AND COLLECTIVELY,
-      "RELEASED PARTIES") BE LIABLE FOR ANY CLAIMS, DEMANDS, LIABILITIES, COSTS,
-      LOSSES, DAMAGES OR EXPENSES (INCLUDING WITHOUT LIMITATION LEGAL COSTS AND
-      ATTORNEYS’ FEES) CAUSED TO OR SUFFERED BY ANY PERSONS OR ENTITIES
-      (INCLUDING WITHOUT LIMITATION, YOU), THAT DIRECTLY OR INDIRECTLY ARISE OUT
-      OF OR RESULT FROM THE USE OF, THE INABILITY TO USE, OR ERRORS, OMISSIONS,
-      INTERRUPTIONS OR OTHER INACCURACIES IN, THE MONARCH TOOLS OR VIOLATION OF
-      ANY RIGHTS OF ANY THIRD PARTY. WITHOUT LIMITING THE FOREGOING, THE
-      RELEASED PARTIES SHALL NOT BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-      CONSEQUENTIAL, SPECIAL, PUNITIVE OR EXEMPLARY DAMAGES, OR ANY OTHER
-      DAMAGES OF ANY KIND, INCLUDING WITHOUT LIMITATION, DAMAGES FOR LOSS OF
-      BUSINESS, LOSS OF DATA, LOST PROFITS OR BUSINESS INTERRUPTION, UNDER ANY
-      CONTRACT, NEGLIGENCE, STRICT LIABILITY OR OTHER THEORY, EVEN IF A RELEASED
-      PARTY HAS BEEN ADVISED OF THE POSSIBILITY OF SUCH DAMAGES. THIS LIMITATION
-      OF LIABILITY APPLIES TO THE FULLEST EXTENT POSSIBLE.
-    </p>
-  </AppSection>
-
-  <AppSection>
-    <AppHeading>Variant interpretation</AppHeading>
-    <p>
-      The Monarch Initiative sources variant interpretations as provided by the
-      original source, with no consolidation of results, and always has a link
-      out to the original reference. While we display the variants and
-      interpretations as originally reported in the variant databases, some
-      variants in the same genes have been reported on different transcripts.
-      Whenever possible, Monarch will attempt to map all variants in a
-      particular gene to a common transcript to aid in viewing and organizing
-      the variants. However, to aid in variant searches, we will also continue
-      to report the original mapping information. All interpretations were
-      performed by the original submitters and have not been modified in any
-      way. As such, there might be discordant interpretations for the same
-      variant from different archives. For further information or to request a
-      review of the variant, please contact the submitter of the original
-      reference. Variant interpretations are not static as they are based on the
-      current understanding of the variant at the time it was observed in a
-      patient or in the literature. Any reported interpretation may not be
-      up-to-date, and some information may not be reported. Many interpretations
-      will change over time as more information about the variant becomes
-      available. If you have any questions about a variant interpretation or
-      would like use any variant interpretations for publication purposes,
-      please contact the submitter of the original reference. As always, we
-      value your feedback. Please email us if you have any problems, concerns,
-      or suggestions.
-    </p>
-  </AppSection>
-</template>
