--- frontend/src/pages/explore/tabs.json
@@ -1,26 +1,0 @@
-[
-  {
-    "id": "search",
-    "text": "Search Graph",
-    "icon": "magnifying-glass",
-    "tooltip": "Search our knowledge graph for genes, diseases, phenotypes, and more, then see rich information and relational data between them"
-  },
-  {
-    "id": "text-annotator",
-    "text": "Text Annotator",
-    "icon": "subscript",
-    "tooltip": "Annotate text with concepts from our knowledge graph"
-  },
-  {
-    "id": "phenotype-explorer",
-    "text": "Phenotype Explorer",
-    "icon": "bars-progress",
-    "tooltip": "Construct two sets of phenotypes and see comparisons between them"
-  },
-  {
-    "id": "resources",
-    "text": "Resources",
-    "icon": "info",
-    "tooltip": "View metadata about the knowledge graph"
-  }
-]
