--- frontend/src/pages/explore/text-annotator.json
@@ -1,3 +1,0 @@
-{
-  "content": "Lewis (1978) found 7 affected males and 5 affected females in 3 consecutive generations of a Caucasian kindred. As in the X-linked Nettleship-Falls form of ocular albinism (300500), the patients showed reduced visual acuity, photophobia, nystagmus, translucent irides, strabismus, hypermetropic refractive errors, and albinotic fundus with foveal hypoplasia. The skin lesions showed macromelanosomes as in X-linked ocular albinism. Deafness, which was accompanied by vestibular hypofunction, lentigines even in unexposed areas, optic nerve dysplasia, and dominant inheritance distinguished this form of ocular albinism. (In the LEOPARD syndrome (151100) vestibular function is normal.)\n\nBard (1978) described a kindred that was atypical of Waardenburg syndrome (see 193510) in several ways. Although the nasal root was prominent, one affected person had dystopia of the inner canthi or lower puncta. The face in some showed striking freckling of pale skin. Symptomatic vestibular disturbance was another unusual feature. Lewis (1989) expressed the opinion that the family reported by Bard (1978) as an instance of Waardenburg syndrome in fact had this disorder. Lewis (1989) had also been told of 2 other small families with the syndrome. Goldberg (1966) described a Waardenburg syndrome family with apparent ocular albinism.\n\nMorell et al. (1997) presented an update of the clinical findings in the family of Bard (1978). The deafness was sensorineural and congenital. Heterochromia iridis was a prominent feature in 1 sibship in which both segmental iris bicolor and complete heterochromia occurred. Most of the affected individuals showed transillumination defects of the iris. Hypopigmentation of the fundus was mild in some, moderate in others, and severe in yet others. Almost all affected individuals had strabismus and visual acuity defects. One individual with a prominent white forelock, characteristic of Waardenburg syndrome, was pictured."
-}
