--- frontend/src/pages/help/PageFeedback.vue
@@ -1,16 +1,0 @@
-<!--
-  help feedback page
-
-  a dedicated page/url for the same feedback form that is openable from any page
--->
-
-<template>
-  <AppSection>
-    <AppHeading>Feedback Form</AppHeading>
-    <TheFeedbackForm />
-  </AppSection>
-</template>
-
-<script setup lang="ts">
-import TheFeedbackForm from "@/components/TheFeedbackForm.vue";
-</script>
