+++ frontend/src/router/descriptions.json
@@ -1,15 +1,28 @@
 {
-  "Explore": "Explore our rich knowledge graph and analysis tools.",
-  "About": "Learn more about what Monarch is, how it works, and who's behind it.",
-  "Help": "Request a feature, report a bug, or chat with us about anything Monarch-related.",
-
-  "Overview": "How all the pieces of Monarch fit together.",
-  "API": "The API behind this website.",
-  "Sources": "Datasets, ontologies, and downloads.",
-  "Cite": "How to cite and attribute Monarch.",
-  "Team": "The people behind Monarch.",
-  "Publications": "Monarch-related published works.",
-  "Terms": "Disclaimers and agreements.",
-
-  "Feedback": "Tell us how we can improve Monarch."
+  "Home": "Explore the Monarch knowledge graph homepage.",
+  "Node": "Details for a single entity in the knowledge graph.",
+  "Testbed": "Experimental playground for in-progress features.",
+  "NotFound": "The page you requested could not be found.",
+  "Phenotype Similarity Search": "Find diseases or genes by comparing phenotype profiles.",
+  "Text Annotator": "Extract ontology terms from unstructured text.",
+  "About KG": "What the Monarch Knowledge Graph is and how it’s built.",
+  "Citation": "How to cite Monarch and attribute data sources.",
+  "How to & Help": "Guides, FAQs, and support for using Monarch.",
+  "Status & QC": "Operational status and quality control of the KG.",
+  "KG Dashboard": "Key metrics, coverage, and trends across the KG.",
+  "Search Results": "Results for your Monarch search query.",
+  "Documentation": "Technical docs and API references for Monarch.",
+  "Terms of Use": "Legal terms, disclaimers, and permitted use.",
+  "Downloads": "Downloads for the knowledge graph available in TSV format, ",
+  "Sources": "List of primary and aggreator sources for the knowledge graph.",
+  "Our Story": "Background, mission, and the core values of Monarch.",
+  "Team": "Meet the people building and maintaining Monarch.",
+  "Scientfic Advisory Board Members": "Members of the Scientific Advisory Board.",
+  "Contact Us": "Ways to reach the Monarch team.",
+  "Funding": "Funding sources and acknowledgments.",
+  "Publications": "Peer-reviewed publications related to Monarch.",
+  "Get Involved": "Community, contributions, and collaboration opportunities.",
+  "OntologyPage": "Ontology resource details and metadata.",
+  "RegistryPage": "Registry resource details and metadata.",
+  "ToolPage": "Tool resource details and metadata."
 }
